update Fines.java
Passed all large test. 
Time complexity O(nlogn).



Implemented merge sort to sort priority in descending order; method merge sort recursively break down array to standalone number and recursively call method merge to sort the array. 

Merge method: two array left and right holds the value of current array to be sorted. Loop through the array in order to add it to priority[]. 3 variables leftIndex, rightIndex and subIndex maintain the current index of the left, right and priority array respectively. 

Start the loop by comparing leftArray[0] with rightArray[0]. If a smaller priority is happened to be placed at the left array; incremented count by the amount of items that was not compared in the left array then add rightArray[0] to the priority array. Repeat this step until either leftIndex or rightIndex surpasses the array length. 
Once either left or right array has been fully added to priority; move on to add the rest of the sorted array to priority.